+++ src/main/java/com/example/trx/apis/event/dto/request/CreateContestEventRequest.java
@@ -1,4 +1,4 @@
-package com.example.trx.apis.event.dto;
+package com.example.trx.apis.event.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
