+++ src/main/java/com/example/trx/apis/event/dto/request/EditScoreRequest.java
@@ -1,4 +1,4 @@
-package com.example.trx.apis.event.dto;
+package com.example.trx.apis.event.dto.request;
 
 import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
@@ -6,8 +6,10 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class SubmitScoreRequest {
+public class EditScoreRequest {
   private Long judgeId;
   private BigDecimal scoreTotal;
   private String breakdownJson;
+  private String editedBy;
+  private String editReason;
 }
