+++ src/main/java/com/example/trx/apis/event/dto/request/SubmitScoreRequest.java
@@ -1,4 +1,4 @@
-package com.example.trx.apis.event.dto;
+package com.example.trx.apis.event.dto.request;
 
 import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
@@ -6,10 +6,8 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class EditScoreRequest {
+public class SubmitScoreRequest {
   private Long judgeId;
   private BigDecimal scoreTotal;
   private String breakdownJson;
-  private String editedBy;
-  private String editReason;
 }
