+++ src/main/java/com/example/trx/apis/event/dto/response/RunResponse.java
@@ -1,4 +1,4 @@
-package com.example.trx.apis.event.dto;
+package com.example.trx.apis.event.dto.response;
 
 import java.util.List;
 import lombok.Builder;
